+++ example/lib/src/file_picker_demo.dart
@@ -103,11 +103,15 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
           allowedExtensions: _allowedExtensionsFromInput(),
           dialogTitle: _dialogTitleController.text,
           initialDirectory: _initialDirectoryController.text,
-          windowsOptions: WindowsOptions(lockParentWindow: _lockParentWindow),
+          windowsOptions: WindowsOptions(
+            lockParentWindow: _lockParentWindow,
+            acceptLabel: _acceptLabelFromInput(),
+          ),
           linuxOptions: LinuxOptions(
             lockParentWindow: _lockParentWindow,
             acceptLabel: _acceptLabelFromInput(),
           ),
+          darwinOptions: DarwinOptions(acceptLabel: _acceptLabelFromInput()),
           androidOptions:
               _androidSafOptionsFromFlags() ?? const AndroidOptions(),
         );
@@ -119,11 +123,15 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
           allowedExtensions: _allowedExtensionsFromInput(),
           dialogTitle: _dialogTitleController.text,
           initialDirectory: _initialDirectoryController.text,
-          windowsOptions: WindowsOptions(lockParentWindow: _lockParentWindow),
+          windowsOptions: WindowsOptions(
+            lockParentWindow: _lockParentWindow,
+            acceptLabel: _acceptLabelFromInput(),
+          ),
           linuxOptions: LinuxOptions(
             lockParentWindow: _lockParentWindow,
             acceptLabel: _acceptLabelFromInput(),
           ),
+          darwinOptions: DarwinOptions(acceptLabel: _acceptLabelFromInput()),
           androidOptions:
               _androidSafOptionsFromFlags() ?? const AndroidOptions(),
         );
@@ -566,7 +574,7 @@ class _FilePickerDemoState extends State<FilePickerDemo> {
         child: TextField(
           decoration: const InputDecoration(
             border: OutlineInputBorder(),
-            labelText: 'Accept Label (Linux)',
+            labelText: 'Accept Label (Linux/Windows/macOS)',
           ),
           controller: _acceptLabelController,
         ),